be/requirements.txt
fastapi==0.86.0
uvicorn==0.20.0
mysql-connector-python==8.0.28
pydantic==1.10.2
passlib==1.7.4
secrets==1.0.2
jwt==1.3.1
bcrypt==4.0.1